+++ fractal_server/app/routes/admin/v2/task.py
@@ -2,11 +2,15 @@ from typing import Any
 
 from fastapi import APIRouter
 from fastapi import Depends
+from fastapi import HTTPException
+from fastapi import Response
+from fastapi import status
 from pydantic import BaseModel
 from pydantic import EmailStr
 from pydantic import Field
 from sqlmodel import func
 from sqlmodel import select
+from sqlmodel import update
 
 from fractal_server.app.db import AsyncSession
 from fractal_server.app.db import get_async_db
@@ -22,6 +26,11 @@ from fractal_server.app.routes.pagination import PaginationResponse
 from fractal_server.app.routes.pagination import get_paginated_response
 from fractal_server.app.routes.pagination import get_pagination_params
 from fractal_server.app.schemas.v2.task import TaskType
+from fractal_server.types import ListUniqueNonNegativeInt
+
+from ._aux_functions_core_tasks import (
+    _verify_non_duplication_task_core_constraint,
+)
 
 router = APIRouter()
 
@@ -30,6 +39,7 @@ class TaskMinimal(BaseModel):
     id: int
     name: str
     type: str
+    is_core: bool
     taskgroupv2_id: int
     command_non_parallel: str | None = None
     command_parallel: str | None = None
@@ -64,6 +74,7 @@ async def query_tasks(
     modality: str | None = None,
     author: str | None = None,
     resource_id: int | None = None,
+    only_core: bool = False,
     pagination: PaginationRequest = Depends(get_pagination_params),
     user: UserOAuth = Depends(current_superuser_act),
     db: AsyncSession = Depends(get_async_db),
@@ -106,6 +117,9 @@ async def query_tasks(
         stm_count = stm_count.join(
             TaskGroupV2, TaskGroupV2.id == TaskV2.taskgroupv2_id
         ).where(TaskGroupV2.resource_id == resource_id)
+    if only_core is True:
+        stm = stm.where(TaskV2.is_core)
+        stm_count = stm_count.where(TaskV2.is_core)
 
     response = await get_paginated_response(
         stm=stm, stm_count=stm_count, pagination=pagination, db=db
@@ -161,3 +175,101 @@ async def query_tasks(
         current_page=response.current_page,
         items=task_info_list,
     )
+
+
+@router.post(
+    "/make-core/",
+    status_code=status.HTTP_200_OK,
+)
+async def make_task_core(
+    task_ids: ListUniqueNonNegativeInt,
+    superuser: UserOAuth = Depends(current_superuser_act),
+    db: AsyncSession = Depends(get_async_db),
+) -> Response:
+    res = await db.execute(
+        select(TaskV2, TaskGroupV2)
+        .join(TaskGroupV2, TaskGroupV2.id == TaskV2.taskgroupv2_id)
+        .where(TaskV2.id.in_(task_ids))
+    )
+    tasks_and_groups = res.all()
+    if len(tasks_and_groups) != len(task_ids):
+        missing_ids = sorted(
+            list(set(task_ids) - set([tg[0].id for tg in tasks_and_groups]))
+        )
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Not all tasks were found (Missing IDs: {missing_ids}).",
+        )
+
+    # Acquire a lock on all rows that could result into conflicting core tasks,
+    # to avoid a race condition where two "make-core" endpoints are called at
+    # the same time. See
+    # https://www.postgresql.org/docs/current/sql-select.html#SQL-FOR-UPDATE-SHARE
+    # and https://www.postgresql.org/docs/current/explicit-locking.html#LOCKING-ROWS
+    await db.execute(
+        select(TaskV2)
+        .where(TaskV2.name.in_([t.name for t, _ in tasks_and_groups]))
+        .where(TaskV2.version.in_([t.version for t, _ in tasks_and_groups]))
+        .where(TaskV2.is_core.is_(False))
+        .with_for_update()
+    )
+
+    payload_tuples = [
+        (
+            task.name,
+            task_group.pkg_name,
+            task_group.version,
+            task_group.resource_id,
+        )
+        for task, task_group in tasks_and_groups
+    ]
+    if len(set(payload_tuples)) != len(payload_tuples):
+        raise HTTPException(
+            status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
+            detail=(
+                "This request would generate conflicting core tasks "
+                "(with the same task name and task-group properties). "
+                "Hint: include fewer tasks in the request body and retry."
+            ),
+        )
+
+    # Non-duplication check constraint
+    for task, task_group in tasks_and_groups:
+        await _verify_non_duplication_task_core_constraint(
+            task=task, task_group=task_group, db=db
+        )
+
+    # Update
+    await db.execute(
+        update(TaskV2).where(TaskV2.id.in_(task_ids)).values(is_core=True)
+    )
+    await db.commit()
+
+    return Response(status_code=status.HTTP_200_OK)
+
+
+@router.post(
+    "/make-not-core/",
+    status_code=status.HTTP_200_OK,
+)
+async def make_task_not_core(
+    task_ids: ListUniqueNonNegativeInt,
+    superuser: UserOAuth = Depends(current_superuser_act),
+    db: AsyncSession = Depends(get_async_db),
+) -> Response:
+    res = await db.execute(select(TaskV2).where(TaskV2.id.in_(task_ids)))
+    tasks = res.scalars().all()
+    if len(tasks) != len(task_ids):
+        missing_ids = sorted(list(set(task_ids) - set([t.id for t in tasks])))
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Not all tasks were found (Missing IDs: {missing_ids}).",
+        )
+
+    # Update
+    await db.execute(
+        update(TaskV2).where(TaskV2.id.in_(task_ids)).values(is_core=False)
+    )
+    await db.commit()
+
+    return Response(status_code=status.HTTP_200_OK)